requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
cairosvg>=2.7.0
PyGObject>=3.42.0; sys_platform == 'linux'

